Removed utility package from project structure and created new mapper package Moved documentType to now live in document package
