+++ document-generator-api/src/main/java/uk/gov/companieshouse/document/generator/api/service/DocumentTypeService.java
@@ -1,7 +1,7 @@
 package uk.gov.companieshouse.document.generator.api.service;
 
 import uk.gov.companieshouse.document.generator.api.exception.DocumentGeneratorServiceException;
-import uk.gov.companieshouse.document.generator.api.utility.DocumentType;
+import uk.gov.companieshouse.document.generator.api.document.DocumentType;
 
 public interface DocumentTypeService {
