+++ document-generator-api/src/main/java/uk/gov/companieshouse/document/generator/api/service/impl/DocumentTypeServiceImpl.java
@@ -3,7 +3,7 @@ package uk.gov.companieshouse.document.generator.api.service.impl;
 import org.springframework.stereotype.Service;
 import uk.gov.companieshouse.document.generator.api.exception.DocumentGeneratorServiceException;
 import uk.gov.companieshouse.document.generator.api.service.DocumentTypeService;
-import uk.gov.companieshouse.document.generator.api.utility.DocumentType;
+import uk.gov.companieshouse.document.generator.api.document.DocumentType;
 
 import java.util.Arrays;
